Updating include file

+++ include/bgezdb.h
@@ -6,8 +6,6 @@
 #include <vector>
 #include <unordered_map>
 #include <cstring>
-#include <mutex>
-#include <variant>
 
 #include "sqlite3.h"
 #include "crow.h"
@@ -16,70 +14,25 @@
 #include "bgezuser.h"
 #include "bgezitem.h"
 
-using DBArg = std::variant<int, std::string, double>;
-using DBArgList = std::vector<DBArg>;
 
-class DBConnection{                               //List of functions for interacting with the DB.
-private:
-    inline static std::mutex mtx{};
-    sqlite3* db;
-
-    bool bindValue(sqlite3_stmt* stmt, int index, int arg){
-        return sqlite3_bind_int(stmt, index, arg) == SQLITE_OK;
-    }
-
-    bool bindValue(sqlite3_stmt* stmt, int index, double arg){
-        return sqlite3_bind_double(stmt, index, arg) == SQLITE_OK;
-    }
-
-    bool bindValue(sqlite3_stmt* stmt, int index, const std::string& arg){
-        return sqlite3_bind_text(stmt, index, arg.c_str(), -1, SQLITE_TRANSIENT) == SQLITE_OK;
-    }
-
-    bool bindValue(sqlite3_stmt* stmt, int index, const DBArg& arg){
-        return std::visit([&](auto&& value){
-            return bindValue(stmt, index, value);
-            }, arg);
-    }
-public:
-    DBConnection(){
+namespace DBCore{                               //List of functions for interacting with the DB.
+    template <typename T>
+    inline bool accessDB(const std::string& query, T get){
+        sqlite3* db{};
         if(sqlite3_open("core.db", &db) == SQLITE_OK){
             CROW_LOG_DEBUG << "DB Opened.";
-        } else {
-            CROW_LOG_DEBUG << "DB failed to open.";
-            sqlite3_close(db);
-        }
-    }
-
-    ~DBConnection(){
-        if (db != nullptr){
-            sqlite3_close(db);
-            db = nullptr;
-        }
-    }
-
-    template <typename T>
-    inline bool accessDB(const std::string& query, const std::vector<DBArgList>& qArgs, T get){
-        std::lock_guard<std::mutex> _(mtx);
-        sqlite3_stmt* statement{};
-        const char* tail{};
-        CROW_LOG_DEBUG << query << "\n";
-        const char* st{query.c_str()};
-        int prepstmt{sqlite3_prepare_v2(db, st, -1, &statement, &tail)};
-        if (prepstmt != SQLITE_OK){
-            CROW_LOG_ERROR << "Failed to prepare statement: " << sqlite3_errmsg(db) << "\n";
-            if (statement != nullptr){
-                sqlite3_finalize(statement);
-            };
-            return false;
-        }
-        for(int j = 0; j < static_cast<int>(qArgs.size()); ++j){
-            for(int i = 0; i < static_cast<int>(qArgs[j].size()); ++i){
-                if (!bindValue(statement,i+1, qArgs[j][i])){
-                    CROW_LOG_ERROR << "Failed to bind Statement. Aborting Query.";
+            sqlite3_stmt* statement{};
+            const char* tail{};
+            CROW_LOG_DEBUG << query << "\n";
+            const char* st{query.c_str()};
+            int prepstmt{sqlite3_prepare_v2(db, st, -1, &statement, &tail)};
+            if (prepstmt != SQLITE_OK){
+                CROW_LOG_ERROR << "Failed to prepare statement: " << sqlite3_errmsg(db) << "\n";
+                if (statement != nullptr){
                     sqlite3_finalize(statement);
-                    return false;
-                }
+                };
+                sqlite3_close(db);
+                return false;
             }
             CROW_LOG_DEBUG << "DB statement prepared.";
             int res{};
@@ -91,68 +44,29 @@ public:
                 } else {
                     CROW_LOG_ERROR << "Bad step. Code: " << res << "\n";
                     sqlite3_finalize(statement);
+                    sqlite3_close(db);
                     return false;
                 }
             }
             if (res == SQLITE_DONE){
                 CROW_LOG_INFO << "Query Complete!\n";
             }
-            sqlite3_reset(statement);
+            sqlite3_finalize(statement);
+            sqlite3_close(db);
+            return true;
         }
-        sqlite3_finalize(statement);
-        return true;
+        return false;
     }
-    template<typename T>
-    bool accessDB(const std::string& query, const std::vector<DBArg>& qArgs, T get){
-        std::lock_guard<std::mutex> _(mtx);
-        sqlite3_stmt* statement{};
-        const char* tail{};
-        CROW_LOG_DEBUG << query << "\n";
-        const char* st{query.c_str()};
-        int prepstmt{sqlite3_prepare_v2(db, st, -1, &statement, &tail)};
-        if (prepstmt != SQLITE_OK){
-            CROW_LOG_ERROR << "Failed to prepare statement: " << sqlite3_errmsg(db) << "\n";
-            if (statement != nullptr){
-                sqlite3_finalize(statement);
-            };
-            return false;
-        }
-        for(int i = 0; i < static_cast<int>(qArgs.size()); ++i){
-            if (!bindValue(statement,i+1, qArgs[i])){
-                CROW_LOG_ERROR << "Failed to bind Statement. Aborting Query.";
-                sqlite3_finalize(statement);
-                return false;
-            }
-        }
-
-        CROW_LOG_DEBUG << "DB statement prepared.";
-        int res{};
-        while((res = sqlite3_step(statement)) != SQLITE_DONE){
-            CROW_LOG_DEBUG << res;
-            if (res == SQLITE_ROW){
-                get(statement);
-                CROW_LOG_DEBUG << "Statement Step.";
-            } else {
-                CROW_LOG_ERROR << "Bad step. Code: " << res << "\n";
-                sqlite3_finalize(statement);
-                return false;
-            }
-        }
-        if (res == SQLITE_DONE){
-            CROW_LOG_INFO << "Query Complete!\n";
-        }
-        sqlite3_finalize(statement);
-        return true;
-    }
-};
-
-namespace DBCore{
-    inline User getUser(const std::string& name, DBConnection& con){      //getUser: retrieves user information from database.
+    inline User getUser(const std::string& name){      //getUser: retrieves user information from database.
+        using namespace std::literals::string_literals;
         CROW_LOG_DEBUG << name;
         User user{};
-        std::string s{"SELECT * FROM Users WHERE username = ?;"};
-        DBArgList arg{name};
-        con.accessDB(s, arg, [&user](sqlite3_stmt* statement){
+        std::string s{
+            ("SELECT * "s)+
+            ("FROM Users "s)+
+            ("WHERE username = '"s)+name+("';"s)
+        };
+        accessDB(s, [&user](sqlite3_stmt* statement){
             const int id{sqlite3_column_int(statement, 0)};
             const std::string uname{reinterpret_cast<const char*>(sqlite3_column_text(statement, 1))};
             const std::string passw{reinterpret_cast<const char*>(sqlite3_column_text(statement, 2))};
@@ -163,12 +77,16 @@ namespace DBCore{
         return user;
     }
 
-    inline User getUser(const int id, DBConnection& con){      //getUser: retrieves user information from database.
+    inline User getUser(const int id){      //getUser: retrieves user information from database.
+        using namespace std::literals::string_literals;
         CROW_LOG_DEBUG << id;
         User user{};
-        std::string s{"SELECT * FROM Users WHERE uid = ?;"};
-        DBArgList arg{id};
-        con.accessDB(s, arg, [&user](sqlite3_stmt* statement){
+        std::string s{
+            ("SELECT * "s)+
+            ("FROM Users "s)+
+            ("WHERE uid = "s)+std::to_string(id)+(";"s)
+        };
+        accessDB(s, [&user](sqlite3_stmt* statement){
             const int id{sqlite3_column_int(statement, 0)};
             const std::string uname{reinterpret_cast<const char*>(sqlite3_column_text(statement, 1))};
             const std::string passw{reinterpret_cast<const char*>(sqlite3_column_text(statement, 2))};
@@ -179,36 +97,45 @@ namespace DBCore{
         return user;
     }
 
-    inline bool addUser(const User& user, DBConnection& con){
-        std::string s{"INSERT INTO Users VALUES (?, ?, ?, ?, ?); "};
-        DBArgList args{
-            user.uid(),
-            user.name(),
-            user.pass(),
-            user.email(),
-            static_cast<int>(user.pref())
+    inline bool addUser(const User& user){
+    using namespace std::literals::string_literals;
+        std::string s{
+            ("INSERT INTO Users VALUES ("s) +
+            std::to_string(user.uid()) + ", '"s +
+            user.name() + "', '"s +
+            user.pass() + "', '"s +
+            user.email() + "',"s +
+            std::to_string(user.pref()) + ");"s
         };
-        return con.accessDB(s, args, [](sqlite3_stmt*){});
+        return accessDB(s, [](sqlite3_stmt*){});
     }
 
-    inline bool deleteUser(const User& user, DBConnection& con){
-        std::string s{"DELETE FROM Users WHERE uid = ?;"};
-        DBArgList arg{user.uid()};
-        return con.accessDB(s, arg, [](sqlite3_stmt*){});
+    inline bool deleteUser(const User& user){
+    using namespace std::literals::string_literals;
+        std::string s{("DELETE FROM Users WHERE uid = '"s) + std::to_string(user.uid()) + "';"s};
+        return accessDB(s, [](sqlite3_stmt*){});
     }
 
-    inline bool addItem(const User& user, const Item& item, DBConnection& con){
+    inline bool addItem(const User& user, const Item& item){
         CROW_LOG_DEBUG << "Adding item to db.";
-        std::string s{"INSERT INTO UserItems VALUES (?, ?);"};
-        DBArgList args{user.uid(), item.id()};
-        return con.accessDB(s, args, [](sqlite3_stmt*){});
+    using namespace std::literals::string_literals;
+        std::string s{
+            ("INSERT INTO UserItems VALUES ("s) +
+            std::to_string(user.uid()) + ", "s +
+            std::to_string(item.id()) + ");"s
+        };
+        return accessDB(s, [](sqlite3_stmt*){});
     }
 
-    inline Item getItem(const std::string& name, DBConnection& con){
+    inline Item getItem(const std::string& name){
+    using namespace std::literals::string_literals;
         Item item{};
-        std::string s{"SELECT * FROM Ingredients WHERE name = ?;"};
-        DBArgList arg{name};
-        con.accessDB(s, arg, [&item](sqlite3_stmt* statement){
+        std::string s{
+            ("SELECT * "s)+
+            ("FROM Ingredients "s)+
+            ("WHERE name = '"s)+name+("';"s)
+        };
+        accessDB(s, [&item](sqlite3_stmt* statement){
             const int id{sqlite3_column_int(statement, 0)};
             const std::string name{reinterpret_cast<const char*>(sqlite3_column_text(statement, 1))};
             const std::string desc{reinterpret_cast<const char*>(sqlite3_column_text(statement, 2))};
@@ -219,11 +146,15 @@ namespace DBCore{
         return item;
     }
 
-    inline Item getItem(int id, DBConnection& con){
+    inline Item getItem(int id){
+    using namespace std::literals::string_literals;
         Item item{};
-        std::string s{"SELECT * FROM Ingredients WHERE id = ?;"};
-        DBArgList arg{id};
-        con.accessDB(s, arg, [&item](sqlite3_stmt* statement){
+        std::string s{
+            ("SELECT * "s)+
+            ("FROM Ingredients "s)+
+            ("WHERE id = '"s)+std::to_string(id)+("';"s)
+        };
+        accessDB(s, [&item](sqlite3_stmt* statement){
             const int id{sqlite3_column_int(statement, 0)};
             const std::string name{reinterpret_cast<const char*>(sqlite3_column_text(statement, 1))};
             const std::string desc{reinterpret_cast<const char*>(sqlite3_column_text(statement, 2))};
@@ -234,17 +165,21 @@ namespace DBCore{
         return item;
     }
 
-    inline crow::json::wvalue getItemList(const User& user, DBConnection& con){
+    inline crow::json::wvalue getItemList(const User& user){
+    using namespace std::literals::string_literals;
         std::vector<int> ids{};
-        std::string s{"SELECT iid FROM UserItems WHERE uid = ?;"};
-        DBArgList arg{user.uid()};
-        con.accessDB(s, arg, [&ids](sqlite3_stmt* statement){
+        std::string s{
+            ("SELECT iid "s)+
+            ("FROM UserItems "s)+
+            ("WHERE uid = "s)+std::to_string(user.uid())+(";"s)
+        };
+        accessDB(s, [&ids](sqlite3_stmt* statement){
             ids.push_back(sqlite3_column_int(statement, 0));
         });
 
         std::vector<Item> items{};
         for (int i : ids){
-            items.push_back(getItem(i, con));
+            items.push_back(getItem(i));
         }
 
         crow::json::wvalue res;
@@ -259,12 +194,11 @@ namespace DBCore{
         return res;
     }
 
-    inline bool deleteItem(const Item& item, DBConnection& con){
-        std::string s{"DELETE FROM UserItems WHERE iid = ?;"};
-        DBArgList arg{item.id()};
-        return con.accessDB(s, arg, [](sqlite3_stmt*){});
+    inline bool deleteItem(const Item& item){
+        using namespace std::literals::string_literals;
+        std::string s{"DELETE FROM UserItems WHERE iid = "s + std::to_string(item.id() + ";"s};
+        return accessDB(s, [](sqlite3_stmt*){});
     }
-
 }
 
 #endif
